Move reflections sorting logic to hook
